Update files { PermutationInt32.cs, PermutationInt64.cs }.
- Unify magic numbers.
- Fix data types.

+++ DotNetTransformer/Math/Group/Permutation/PermutationInt64.cs
@@ -45,7 +45,7 @@ namespace DotNetTransformer.Math.Group.Permutation {
 				}
 				if(multFlag == 1) return 1;
 				if((multFlag & -0x2000) != 0) return ((multFlag >> 14) & 3) + 14;
-				byte r = 1;
+				int r = 1;
 				if((multFlag & 0x0AAA) != 0) r *= 2;
 				if((multFlag & 0x0924) != 0) r *= 3;
 				if((multFlag & 0x0888) != 0) r *= 2;
